Extract user password authenication methods
