+++ app/models/user.rb
@@ -35,15 +35,15 @@
 #  daily_summary_minute              :integer
 #
 
-require 'digest/sha1'
-
 class User < ActiveRecord::Base
   include AlaveteliFeatures::Helpers
   include AlaveteliPro::PhaseCounts
+  include User::Authenication
+
   rolify before_add: :setup_pro_account
   strip_attributes :allow_empty => true
 
-  attr_accessor :password_confirmation, :no_xapian_reindex
+  attr_accessor :no_xapian_reindex
   attr_accessor :entered_otp_code
 
   has_many :info_requests,
@@ -119,8 +119,6 @@ class User < ActiveRecord::Base
 
   validates_presence_of :email, :message => _("Please enter your email address")
   validates_presence_of :name, :message => _("Please enter your name")
-  validates_presence_of :hashed_password, :message => _("Please enter a password")
-  validates_confirmation_of :password, :message => _("Please enter the same password twice")
 
   validates_length_of :about_me,
     :maximum => 500,
@@ -248,11 +246,6 @@ class User < ActiveRecord::Base
     nil # so doesn't print all users on console
   end
 
-  def self.encrypted_password(password, salt)
-    string_to_hash = password + salt # TODO: need to add a secret here too?
-    Digest::SHA1.hexdigest(string_to_hash)
-  end
-
   def self.record_bounce_for_email(email, message)
     user = User.find_user_by_email(email)
     return false if user.nil?
@@ -377,26 +370,6 @@ class User < ActiveRecord::Base
       suffix_num = suffix_num + 1
     end
     write_attribute(:url_name, unique_url_name)
-  end
-
-  # Virtual password attribute, which stores the hashed password, rather than plain text.
-  def password
-    @password
-  end
-
-  def password=(pwd)
-    @password = pwd
-    if pwd.blank?
-      self.hashed_password = nil
-      return
-    end
-    create_new_salt
-    self.hashed_password = User.encrypted_password(password, salt)
-  end
-
-  def has_this_password?(password)
-    expected_password = User.encrypted_password(password, salt)
-    hashed_password == expected_password
   end
 
   def otp_enabled?
@@ -648,10 +621,6 @@ class User < ActiveRecord::Base
   end
 
   private
-
-  def create_new_salt
-    self.salt = object_id.to_s + rand.to_s
-  end
 
   def set_defaults
     if new_record?
